# runtime/runc: selectively lock events to reduce contention issues