### Context
See: https://github.com/containerd/containerd/issues/8557#issuecomment-1568052031 (writeup of the issue)
fixes #8557 

### What I did

Address the performance issue by reducing contention between https://github.com/containerd/containerd/blob/4b7145cfd38d76bc763e3cee1210963bc6c22680/runtime/v2/runc/task/service.go#L154
and
https://github.com/containerd/containerd/blob/4b7145cfd38d76bc763e3cee1210963bc6c22680/runtime/v2/runc/task/service.go#L526
by removing
https://github.com/containerd/containerd/blob/4b7145cfd38d76bc763e3cee1210963bc6c22680/runtime/v2/runc/task/service.go#L104
and instead, individually keep track of starting processes which allows `checkProcesses()` to optimistically check unblocked/already started processes, and only if necessary separately queue events for blocked processes to be retried on a separate goroutine.

--------------------

This model significantly reduces contention around this area, allowing https://github.com/containerd/containerd/blob/4b7145cfd38d76bc763e3cee1210963bc6c22680/runtime/v2/runc/task/service.go#L526
to process the incoming events which decongests the line down to https://github.com/containerd/containerd/blob/4b7145cfd38d76bc763e3cee1210963bc6c22680/sys/reaper/reaper_unix.go#L64

### Benchmarks

(using the repro steps from https://github.com/containerd/containerd/issues/8557#issue-1722590080)

baseline:
```
0.11user 0.03system 0:01.00elapsed
0.09user 0.06system 0:25.40elapsed
0.08user 0.02system 0:17.45elapsed
```

with this change:
```
0.11user 0.08system 0:00.91elapsed
0.08user 0.02system 0:00.72elapsed
0.09user 0.01system 0:00.65elapsed
```

With some instrumenting, we can also confirm that the intended effect is achieved since the notify goroutine:

baseline:
```
notify goroutine took 10.167µs
notify goroutine took 6.542µs
notify goroutine took 16.5µs
notify goroutine took 4.417µs
notify goroutine took 8.063819576s
notify goroutine took 2.060341814s
notify goroutine took 5.061881112s
notify goroutine took 7.063627129s
notify goroutine took 1.060047492s
notify goroutine took 4.061607998s
notify goroutine took 6.063175599s
notify goroutine took 9.064799771s
```

with this change:
```
notify goroutine took 110.280639ms
notify goroutine took 173.13788ms
notify goroutine took 142.423614ms
notify goroutine took 25.899222ms
notify goroutine took 4.356252ms
notify goroutine took 3.834µs
notify goroutine took 2.833µs
notify goroutine took 2.5µs
notify goroutine took 64.833µs
notify goroutine took 2.197334ms
notify goroutine took 78.833µs
notify goroutine took 59.167µs
notify goroutine took 98.583µs
notify goroutine took 47.042µs
notify goroutine took 42.958µs
notify goroutine took 48.916µs
notify goroutine took 69.125µs
notify goroutine took 57.291µs
notify goroutine took 42.333µs
notify goroutine took 33.417µs
notify goroutine took 72.39987ms
```

*callout*: I believe this solution preserves the intended behaviour re: ensuring start events are sent before exit events (however well this worked before), but PTAL/discuss/propose alternate ways of handling the issue.

### (bonus) cute animal

<img width="439" alt="image" src="https://github.com/containerd/containerd/assets/70572044/76cb418d-5bd8-4618-b33e-95a00644f9f9">