@laurazard 2023-05-30
Edit: I was running some more tests and this doesn't completely solve the issue – every now and then, `checkProcesses` still takes an inordinate amount of time, in the following call:
https://github.com/containerd/containerd/blob/c7b9a95bdc8b6ffd99c8c15c69fd3617e5f20333/runtime/v2/runc/task/service.go#L536
which locks everything else out. I've got some more changes to address that that I'll include in a separate commit (tomorrow, I need sleep 😓) but for right now I'll draft this PR)